Agentic-Ai/task2/requirements.txt
langchain
langchain-community
langchain-groq
langgraph
python-dotenv
wikipedia
